# web/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comic Reader API",
    "description": "Self-hostable comic reader. See comic-reader-spec.md.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/admin/auth/config": {
      "get": {
        "tags": [
          "api::auth_config"
        ],
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthConfigView"
                }
              }
            }
          },
          "403": {
            "description": "admin only"
          }
        }
      }
    },
    "/admin/auth/oidc/discover": {
      "post": {
        "tags": [
          "api::auth_config"
        ],
        "operationId": "probe_discovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OidcDiscoverReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OidcDiscoverResp"
                }
              }
            }
          },
          "400": {
            "description": "invalid issuer URL"
          },
          "403": {
            "description": "admin only"
          },
          "502": {
            "description": "discovery doc unreachable / malformed"
          }
        }
      }
    },
...
    "/admin/fs/list": {
      "get": {
        "tags": [
          "api::admin_fs"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Path to list. Omit (or pass empty) to list the root. The server\nwill reject anything that doesn't canonicalise to a descendant of\n`COMIC_LIBRARY_PATH`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResp"
                }
              }
            }
          },
          "400": {
            "description": "path validation failed"
          },
          "403": {
            "description": "admin only, or path is outside library root"
          },
          "404": {
            "description": "path does not exist"
          }
        }
      }
    },
...
    "/admin/settings": {
      "get": {
        "tags": [
          "api::admin_settings"
        ],
        "operationId": "get_all",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          },
          "403": {
            "description": "admin only"
          }
        }
      },
      "patch": {
        "tags": [
          "api::admin_settings"
        ],
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          },
          "400": {
            "description": "validation"
          },
          "403": {
            "description": "admin only"
          }
        }
      }
    },
...
    "/me/saved-views/preview": {
      "post": {
        "tags": [
          "api::saved_views"
        ],
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesListView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthConfigView": {
        "type": "object",
        "required": [
          "auth_mode",
          "oidc",
          "local"
        ],
        "properties": {
          "auth_mode": {
            "type": "string",
            "description": "`'oidc' | 'local' | 'both'`"
          },
          "local": {
            "$ref": "#/components/schemas/LocalConfigView"
          },
          "oidc": {
            "$ref": "#/components/schemas/OidcConfigView"
          }
        }
      },
...
      "Condition": {
        "type": "object",
        "description": "One condition row in a filter DSL. `group_id` always 0 in v1; reserved\nfor nested-group support without a wire-format break.",
        "required": [
          "field",
          "op"
        ],
        "properties": {
          "field": {
            "$ref": "#/components/schemas/Field"
          },
          "group_id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "op": {
            "$ref": "#/components/schemas/Op"
          },
          "value": {
            "description": "Shape varies by `(field, op)`. The compiler validates and rejects\nmismatches with a clear error. See [`super::registry`] for the\nmatrix of allowed combinations."
          }
        }
      },
...
      "DirEntry": {
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "Field": {
        "type": "string",
        "description": "All filterable fields. Per-field metadata (kind, allowed ops, SQL\ncolumn) lives in [`super::registry`]. Adding a field is a two-step:\nadd a variant here + a registry entry. The OpenAPI schema mirrors this\nenum so the M5 field-picker can iterate variants without a separate\nJSON constant.",
        "enum": [
          "library",
          "name",
          "year",
          "volume",
          "total_issues",
          "publisher",
          "imprint",
          "status",
          "age_rating",
          "language_code",
          "created_at",
          "updated_at",
          "genres",
          "tags",
          "writer",
          "penciller",
          "inker",
          "colorist",
          "letterer",
          "cover_artist",
          "editor",
          "translator",
          "read_progress",
          "last_read",
          "read_count"
        ]
      },
      "FilterDsl": {
        "type": "object",
        "required": [
          "match_mode",
          "conditions"
        ],
        "properties": {
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Condition"
            }
          },
          "match_mode": {
            "$ref": "#/components/schemas/MatchMode"
          }
        }
      },
...
      "ListResp": {
        "type": "object",
        "required": [
          "path",
          "root",
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirEntry"
            }
          },
          "path": {
            "type": "string",
            "description": "Canonical absolute path of the listed directory. The picker uses\nthis for the breadcrumb and as the value to drill back up from."
          },
          "root": {
            "type": "string",
            "description": "Canonical absolute path of the configured library root. The\npicker uses this to know when \"up\" should be disabled."
          }
        }
      },
      "LocalConfigView": {
        "type": "object",
        "required": [
          "enabled",
          "registration_open",
          "smtp_configured"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Local mode is enabled (auth_mode = 'local' or 'both')."
          },
          "registration_open": {
            "type": "boolean",
            "description": "Self-serve registration is open. False locks the system to the\nexisting user set + the first-user admin bootstrap."
          },
          "smtp_configured": {
            "type": "boolean",
            "description": "SMTP wired (host non-empty). Without it, email verification + reset\nflows fall back to bypass-on-register."
          }
        }
      },
...
      "MatchMode": {
        "type": "string",
        "enum": [
          "all",
          "any"
        ]
      },
...
      "OidcConfigView": {
        "type": "object",
        "required": [
          "configured",
          "trust_unverified_email"
        ],
        "properties": {
          "client_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "configured": {
            "type": "boolean"
          },
          "issuer": {
            "type": [
              "string",
              "null"
            ]
          },
          "trust_unverified_email": {
            "type": "boolean",
            "description": "True when `COMIC_OIDC_TRUST_UNVERIFIED_EMAIL` is set. Surfaced because\nit materially weakens email-claim trust."
          }
        }
      },
      "OidcDiscoverReq": {
        "type": "object",
        "required": [
          "issuer"
        ],
        "properties": {
          "issuer": {
            "type": "string",
            "description": "The OIDC issuer URL the admin is about to save. Used as the base\nfor `${issuer}/.well-known/openid-configuration`."
          }
        }
      },
      "OidcDiscoverResp": {
        "type": "object",
        "required": [
          "issuer"
        ],
        "properties": {
          "authorization_endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "end_session_endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "issuer": {
            "type": "string"
          },
          "jwks_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "scopes_supported": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Truncated to the most useful fields above. The full doc is large\nand not all of it is admin-interesting; if a future need arises\nthe raw JSON can be added under a separate field."
          },
          "token_endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "userinfo_endpoint": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Op": {
        "type": "string",
        "enum": [
          "contains",
          "starts_with",
          "equals",
          "not_equals",
          "is",
          "is_not",
          "in",
          "not_in",
          "gt",
          "gte",
          "lt",
          "lte",
          "between",
          "before",
          "after",
          "relative",
          "includes_any",
          "includes_all",
          "excludes",
          "is_true",
          "is_false"
        ]
      },
...
      "PreviewReq": {
        "type": "object",
        "required": [
          "filter",
          "sort_field",
          "sort_order",
          "result_limit"
        ],
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/FilterDsl"
          },
          "result_limit": {
            "type": "integer",
            "format": "int32"
          },
          "sort_field": {
            "$ref": "#/components/schemas/SortField"
          },
          "sort_order": {
            "$ref": "#/components/schemas/SortOrder"
          }
        }
      },
...
      "RegistryEntry": {
        "type": "object",
        "required": [
          "key",
          "kind",
          "is_secret"
        ],
        "properties": {
          "is_secret": {
            "type": "boolean"
          },
          "key": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "description": "`\"string\" | \"bool\" | \"uint\" | \"duration\"`."
          }
        }
      },
...
      "ResolvedEntry": {
        "type": "object",
        "required": [
          "key",
          "value",
          "is_secret"
        ],
        "properties": {
          "is_secret": {
            "type": "boolean"
          },
          "key": {
            "type": "string"
          },
          "value": {}
        }
      },
...
      "SeriesListView": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeriesView"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Total matching rows across all pages — populated only on the\nfirst page (no cursor). `None` on subsequent pages where the\ncaller already has this number from page zero, so we don't pay\nthe `COUNT(*)` cost on every page fetch."
          }
        }
      },
...
      "SettingsView": {
        "type": "object",
        "required": [
          "registry",
          "values"
        ],
        "properties": {
          "registry": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistryEntry"
            },
            "description": "Catalog of every DB-overlayable setting recognized by this server\nbuild. Empty in M1; populated milestone-by-milestone."
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResolvedEntry"
            },
            "description": "Current effective values resolved from the DB overlay. Secret\nrows are returned as the literal string `\"<set>\"` so this surface\nnever leaks credentials. Order matches `registry`."
          }
        }
      },
      "SortField": {
        "type": "string",
        "description": "Sort axes a saved view can choose. Distinct from [`Field`] because not\nevery filterable field is a sensible sort axis (e.g., multi-valued\ngenres can't sort).",
        "enum": [
          "name",
          "year",
          "created_at",
          "updated_at",
          "last_read",
          "read_progress"
        ]
      },
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
      "UpdateSettingsReq": {
        "type": "object",
        "additionalProperties": {
          "description": "Map of setting key → new value. A `null` value deletes the row.\nUnknown keys reject the whole batch with a 400."
        }
      },
...
    }
  }
}